backend/pyproject.toml
[project]
name = "longevity-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Jafeth",email = "132289436+JafethHPC@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "fastapi (>=0.121.3,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "requests (>=2.32.5,<3.0.0)",
    "beautifulsoup4 (>=4.14.2,<5.0.0)",
    "biopython (>=1.86,<2.0)",
    "langchain (>=1.0.8,<2.0.0)",
    "langchain-openai (>=1.0.3,<2.0.0)",
    "langchain-community (>=0.4.1,<0.5.0)",
    "langsmith (>=0.4.46,<0.5.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pypdf (>=6.4.1,<7.0.0)",
    "pdf2image (>=1.17.0,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "langchain-text-splitters (>=1.0.0,<2.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "flashrank (>=0.2.10,<0.3.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "httpx (>=0.28.0,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
